Holy code refactoring @marshyonline
