--- server/auth/_helpers.js
@@ -1,12 +1,0 @@
-const bcrypt = require('bcryptjs');
-var db = require('../knex/knex.js');
-var logger = require('../log');
-
-function comparePass(userPassword, databasePassword) {
-    return bcrypt.compareSync(userPassword, databasePassword);
-}
-
-
-module.exports = {
-    comparePass
-};
