+++ server/routes/admin.js
@@ -6,6 +6,7 @@ var fs = require('fs');
 var logger = require('../log');
 var util = require('util');
 var passport = require('../auth/local'); // pass passport for configuration
+const authHelper = require('../middleware/authhelper')
 
 router.use(function (req, res, next) {
     res.locals.login = req.isAuthenticated();
@@ -27,7 +28,7 @@ router.use(bodyParser.urlencoded({     // to support URL-encoded bodies
 }));
 
 router.route('/settingsData')
-    .get(isAdmin, function (req, res, next) {
+    .get(authHelper.isAdmin, function (req, res, next) {
         nconf.load();
         let settings = nconf.get();
         // logger.main.debug(util.format('Config:\n\n%o',settings));
@@ -47,7 +48,7 @@ router.route('/settingsData')
         let data = { "settings": settings, "plugins": plugins, "themes": themes }
         res.json(data);
     })
-    .post(isAdmin, function (req, res, next) {
+    .post(authHelper.isAdmin, function (req, res, next) {
         nconf.load();
         if (req.body) {
             //console.log(req.body);
@@ -61,19 +62,8 @@ router.route('/settingsData')
         }
     });
 
-router.get('*', isAdmin, function (req, res, next) {
+router.get('*', authHelper.isAdminGUI, function (req, res, next) {
     res.render('admin', { pageTitle: 'Admin' });
 });
 
 module.exports = router;
-
-function isAdmin(req, res, next) {
-    // if user is authenticated in the session, carry on 
-    if (req.isAuthenticated() && req.user.role == 'admin') {
-        return next();
-    } else {
-        // if they aren't redirect them to the home page
-        res.status(401).redirect('/');
-        logger.auth.debug(req.user.username + ' attempted to access admin functions')
-    }
-}
